requirements.txt
docopts==0.6.1
requests>=1.2.3
nose>=1.3
